+++ android/src/main/java/xyz/luan/audioplayers/WrappedMediaPlayer.java
@@ -20,7 +20,7 @@ public class WrappedMediaPlayer extends Player implements MediaPlayer.OnPrepared
     private boolean prepared = false;
     private boolean playing = false;
 
-    private double shouldSeekTo = -1;
+    private int shouldSeekTo = -1;
 
     private MediaPlayer player;
     private AudioplayersPlugin ref;
@@ -173,9 +173,9 @@ public class WrappedMediaPlayer extends Player implements MediaPlayer.OnPrepared
     // seek operations cannot be called until after
     // the player is ready.
     @Override
-    void seek(double position) {
+    void seek(int position) {
         if (this.prepared)
-            this.player.seekTo((int) position);
+            this.player.seekTo(position);
         else
             this.shouldSeekTo = position;
     }
@@ -192,7 +192,7 @@ public class WrappedMediaPlayer extends Player implements MediaPlayer.OnPrepared
             ref.handleIsPlaying(this);
         }
         if (this.shouldSeekTo >= 0) {
-            this.player.seekTo((int) this.shouldSeekTo);
+            this.player.seekTo(this.shouldSeekTo);
             this.shouldSeekTo = -1;
         }
     }
